fix(misc): Standardize handling of values bundled with a done message
Iterators may include a defined value in the final iterator result (the
first one in which `done` is `true`), such as when generators terminate
with a `return` statement. Since such values are dropped by all
`for...of` loops, they should also be dropped by the few functions that
manually advance iterators (to wit, `interleave`, `merge`, and `zip`).